requirements.txt
pip==24.2
PyPDF2
Flask
google-auth
google-auth-oauthlib
google-api-python-client
numpy
pandas
scikit-learn
spacy
gunicorn==20.1.0 
gradio


